extract account-decoder-client-types and convert UiAccount::encode to a free func encode_ui_account
