+++ rpc-client/src/rpc_client.rs
@@ -20,8 +20,9 @@ use {
     serde::Serialize,
     serde_json::Value,
     solana_account_decoder::{
+        encode_ui_account,
         parse_token::{UiTokenAccount, UiTokenAmount},
-        UiAccount, UiAccountEncoding,
+        UiAccountEncoding,
     },
     solana_rpc_client_api::{
         client_error::{Error as ClientError, ErrorKind, Result as ClientResult},
@@ -3737,7 +3738,7 @@ pub fn create_rpc_client_mocks() -> crate::mock_sender::Mocks {
                 executable: false,
                 rent_epoch: 0,
             };
-            UiAccount::encode(&pubkey, &account, UiAccountEncoding::Base64, None, None)
+            encode_ui_account(&pubkey, &account, UiAccountEncoding::Base64, None, None)
         },
     })
     .unwrap();
@@ -3995,7 +3996,7 @@ mod tests {
         };
         let keyed_account = RpcKeyedAccount {
             pubkey: pubkey.to_string(),
-            account: UiAccount::encode(&pubkey, &account, UiAccountEncoding::Base64, None, None),
+            account: encode_ui_account(&pubkey, &account, UiAccountEncoding::Base64, None, None),
         };
         let expected_result = vec![(pubkey, account)];
         // Test: without context
